requirements/local.txt
-r base.txt

# Custom extensions for django
django-extensions==1.5.6

# For improved debugging
django-debug-toolbar==1.3.2

# For testing
coverage==4.0b1
flake8==2.4.1
django-test-plus==1.0.8
factory_boy==2.5.2
nose==1.3.7
django-nose==1.4.1
mock-django==0.6.9